+++ client/scripts/views/pages/validators/substrate/index.ts
@@ -1,0 +1,2 @@
+export { default as SubstratePreHeader } from './preheader';
+export { default as SubstratePresentationComponent } from './presentation_component';
